modules/montage: stop assuming modules don't fragment
There's been a longstanding todo item in montage where it was
ignoring the fragmenter returned by a module's prepare_frame().

This commit continues with the single-threaded rendering of the
modules within their respective tiles, still ad-hoc open coded.
But now actually applies the fragmenter returned as if the
rendering were being threaded, since when a module returns a
fragmenter from its prepare_frame() it may strongly depend on
that fragmenting for its output.