til_setup: add til_setup_new() helper
For use when setup functions allocate their private setup to
return in *res_setup.

They specify the size of their private setup, and supply the free
function to use.  This may be libc's free() when it's a simple
setup struct, or a bespoke free function when deep/complex
freeing is required for cleanup.

It's expected that callers will be embedding til_setup_t at the
start of their private setup struct, and returning a pointer to
this in *res_setup which would be the same value as a pointer to
to their private setup struct.