*: free setup allocations via til_setup_free()
This should plug a bulk of the setup leaks.  Some of the
free_funcs still need to be changed to bespoke ones in modules
that allocate nested things in their respective setup, so those
are still leaking the nested things which are usually just a
small strdup of some kind.