modules/rtv: except snow_module=none,snow_duration=0
The code was deliberately showing a single frame of snow when
snow_duration=0, since the durations are integral seconds ther
was no way to have a sub 1-second snow otherwise.

snow_module=none just means cleared for the snow_duration, not no
snow mechanism whatsoever.  So the combination of 0 duration and
"none" was still flashing a single blank frame.

With this commit the combination of "none" and 0 snow_duration
prevents even a single frame of blank snow from being rendered.