til: wire n_cpus up to the fragmenter function
Fragmenting is often dimensioned according to the number of cpus,
and by not supplying this to the fragmenter it was made rather
common for module contexts to plumb this themselves - in some
cases incorporating a context type/create/destroy rigamarole
for the n_cpus circuit alone.

So just plumb it in libtil, and the prepare_frame functions can
choose to ignore it if they have something more desirable onhand.

Future commits will remove a bunch of n_cpus from module contexts
in favor of this.