modules/montage: fix fragnum misuse as cpu #
This was causing the snow module to scribble via montage, since
snow uses per-cpu rand seeds indexed using the cpu value.

I didn't dig in to see if this was a vestigial thing where
fragnum once was passed as a parameter (it's also in
til_fb_fragment_t.number, but probably wasn't always that way).

But it's now used as a cpu idx, but since they're the same type
nothing complained, they say programming is hard.