til: supply a seed to til_module_t.create_context()
In the recent surge of ADD-style rtv+compose focused development,
a bunch of modules were changed to randomize initial states at
context_create() so they wouldn't be so repetitive.

But the way this was done in a way that made it impossible to
suppress the randomized initial state, which sometimes may be
desirable in compositions.  Imagine for instance something like
the checkers module, rendering one module in the odd cells, and
another module into the even cells.  Imagine if these modules are
actually the same, but if checkers used one seed for all the odd
cells and another seed for all the even cells.  If the modules
used actually utilized the seed provided, checkers would be able
to differentiate the odd from even by seeding them differently
even when the modules are the same.

This commit is a step in that direction, but rototiller and all
the composite modules (rtv,compose,montage) are simply passing
rand() as the seeds.  Also none of the modules have yet been
modified to actually make use of these seeds.

Subsequent commits will update modules to seed their
pseudo-randomized initial state from the seed value rather than
always calling things like rand() themselves.