modules/*: first stab at utilizing supplied seeds
This is a mostly mechanical change of using rand_r() in place of
rand(), using the provided seed as the seed state.

There's some outstanding rand()s outside of create_context()
which should probably get switched over, with the seed being
stowed in the context struct.  I didn't bother going deeper on
this at the moment in the interests of getting to sleep soon.