modules/shapes: add spin setting for star+pinwheel
Introduces spin={-1,0,1} with a few intermediate fractions on
both sides of 0.  Controls rate and direction.

As-is these multiple choice style options don't let you
explicitly set a value in rototiller on the commandline that
isn't in the set.

There should probably be a flag in the desc we can set when
bypassing the available options is tolerable, probably when
regex's start getting enforced.

That way rototiller's commandline setting parsing can just lean
on the regex, and if the desc says anything can come in that
passes the regex even if it's not in the values set, this can all
still work and have something resembling input validation.

At the end of the day, the multiple choice value sets are
supposed to be a convenience/guide to a sane variety of values
and of particular utility to randomizers like used by
rtv/montage/compose, but also GUI setting selectors like in
glimmer.

They're not intended to get in the way preventing development
from accessing explicit values of arbitrary precision which can
be really necessary especially when trying to determine what's
best for going in the values set.