modules/shapes: introduce pinch setting
This is kind of experimental, not sure how I feel about it.

pinch=0..1 with a bunch of fractions, 0 disables it.
pinch_spin=-1..1 same as spin=
pinches=1..10  number of pinches, which come in pairs

This applies to all the current shapes, for a tour:
--module=rtv,channels=shapes,duration=2,context_duration=2,snow_module=none --defaults

I think the speeds might go to high atm, I kind of liked the
slower spins before all this more.