til: add --go to supported args
In rototiller this disables the automatic displaying of settings
actually used when they differ from what was explicitly specified
as args.  Which also disables the waiting to press a key.

This should also get used by glimmer to automatically start
rendering without just putting up the configured settings panel
and waiting for a click on "go!".