modules/checkers: experimenting with fill modes
this introduces a color= setting syntax:
 color=#rrggbb
 color=0xrrggbb
 color=rrggbb

 where rrggbb is case-insensitive html-style hexadecimal

also introduces a fill= setting:
 fill=color
 fill=sampled
 fill=textured
 fill=random
 fill=mixed

sampled draws the color from the incoming fragment when layered,
textured draws the pixels from the texture when available,
random randomizes the choice from color,sampled,textured.

mixed isn't implemented fully and is just aliased to random
currently.  The thinking for mixed is to allow specifying
proportions for color,sampled,textured which would then be
applied as weights when randomizing the selection from the three
at every filled checker.

the current implementation is just calling rand() when
randomized, but should really be like the other dynamics in
checkers with rate control and hash-based.

and introduces a fill_module= setting:

this is a first stab at employing other modules for filling the
filled cells.

Note since checkers is already a threaded module, the fill module
context gets created per-cpu but with an n_cpus=1.

This is kind of the first time module contexts are being rendered
manifold for the same frame, and that's illuminating some
shortcomings which needed to be dealt with.  Some modules
automatically advance a phase/T value on every render which gets
persisted in their context struct.  With how checkers is using
contexts, it's desirable for multiple renders of the same context
using the same ticks to produce the same output.  So modules need
to be more careful about time and determine "dt" (delta-time)
values, and animate proportional to ticks elapsed.  When ticks
doesn't change between renders, dt is zero, and nothing should
change.

For now this is using a hard-coded list of modules to choose
from, you specify the module by name or "none" for no
fill_module (solid checker fill).  ex: "fill_module=shapes"

There's a need for something like fragment color and flag
overrides to allow til_module_render() to be treated as more of a
brush where the caller gets to specify what colors to use, or if
texturing should be allowed.  For now, when fill_module=$module
is employed, the color determination stuff within checkers
doesn't get applied.  That will need to be fixed in the future.