til: experimentally fragment-centric page api
It seems like it might be most ergonomic and convenient for
everything to just use til_fb_fragment_t and rely on ops.submit
to determine if the fragment is a page or not, and if it is how
to submit it.

This commit brings things into that state of the world, it feels
kind of gross at the til_fb_page_*() API.  See the large comment
in til_fb.c added by this commit for more information.

I'm probably going to just run with this for now, it can always
get cleaned up later.  What's important is to get the general
snapshotting concept and functionality in place so modules can
make use of it.  There will always be things to cleanup in this
messy tangle of a program.