mem_fb: introduce --video=mem; a dummy in-memory video backend
The immediate impetus for adding this is to enable running
rototiller even on headless machines just for the sake of getting
some FPS measurements.

It'd be nice to get a sense for what FPS rototiller would
experience on larger modern machines like big EPYC or
Threadripper systems.  But it seems most of those I can get
access to via others running them on work hardware or the like
can at most just run it over ssh without any display or risk of
disrupting the physical console.

But this is probably also useful for testing/debugging purposes,
especially since it doesn't bother to synchronize flips on
anything not even a timer.  So a bunch of display complexity is
removed running with video=mem as well as letting the framerate
run unbounded.

Having said that, it might be nice to add an fps=N setting where
mem_fb uses a plain timer for scheduling the flips.

Currently the only setting is size=WxH identical to the sdl_fb
size= setting, defaulting to 640x480.