til_settings: support rudimentary =value escaping
This is a step towards properly handling nested settings, so we
can do stuff like:

--module=rtv,channels=compose\,layers=checkers\\\,fill_module=shapes\\\,size=64\,texture=plasma

and have rtv actually cycle through just compose with
checkers+plasma layers but holding the specified checkers
settings to shapes filler with a size of 64, randomizing the
rest.

There's more work to do before that can actually happen, but
first thing is to just support escaping the settings values.