til_settings: add til_setting_desc_t.as_nested_settings
When this is set, the setting is itself to be a settings instance
that the frontend must create and place in the relevant
til_setting_t.value_as_nested_settings.

This commit implements that frontend portion in
setup_interactively() for the rototiller frontend.

No setup_func() yet attempts to make use of this stuff.  There's
probably more change needed before that can happen, specifically
the setup_func() likely must always produce a til_settings_t* to
indicate which settings instance is currently relevant to the
frontend.  Without setup_func() telling the frontend, the
frontend has basically no other way of knowing when the backend
setup_func() has moved up/down the heirarchy at the current
iteration.