modules/strobe: add rudimentary strobe light module
After reading about the Dreamachine[0], I wanted to experience
this phenomenon.  The javascript-based web implementations
struggled to hold a steady 10Hz rate and would flicker like
crazy, so here we are.

Only setting right now is period=float_seconds, defaults to .1
for 10Hz.

One limitation in the current implementation is when the frame
rate can't keep up with the period the strobe will just stick on
without ever going off, because the period will always be
expired.  There should probably be a setting to force turning off
for at least one frame when it can't keep up.

[0] https://en.wikipedia.org/wiki/Dreamachine