modules/roto: move tables init to context create
Mechanical rearrangement, but ultimately there probably needs to
be an initialize function added to til_module_t.  With all the
threading chaos going on, this approach to implicit
initialization with a static flag is racy without using atomics.

For now it's probably marginally better to do this in context
create vs. prepare frame.  Context creates *tend* to happen in
single-threaded phases of execution, and infrequently.  Prepare
frame is a serialized phase of the rendering for a given context,
but there can be many contexts in-flight simultaneously now with
all the forms of compositing happening, sometimes from multiple
threads.  So that assumption no longer holds...