modules/compose: more robust texture preservation
The existing code only really cared about preserving the incoming
texture on behalf of the caller when the compose code itself was
doing something with the texture.

But there's scenarios where the underlying module being rendered
(or its descendants) might play with the texture, and in such a
situation when the outer compose wasn't involving a texture it'd
let the descandants installed texture leak out to the callers
making the texture apply more than one'd expect.

Arguably none of the modules should be missing restoration of the
incoming texture after installing+rendering with their own.  But
with this change in place, compose will clean up after nested
modules leaking their texture up.