til_module_context: hash the path
The purpose of the context path is to aid in locating the context
instance.  The initial application of this will be in service of
the taps, which require their module's path as a sort of
containing directory of the tap name.  It'd be convenient to
simply add the path hash with the tap hash to produce the tap's
"absolute path" hash when looking up in the hash table.