til_stream: first stab at implementing til_stream_t
Until now there's just been a forward declared type for
til_fb_fragment_t.stream's type, and it's been completely unused.

The purpose of the stream is to provide a continous inter-frame
object where information can be stored pertaining to the stream
of frames.

Right now, that information is limited to emergent "pipes" formed
by using taps against a given stream.  Taps at new paths in the
stream become added as pipes for those paths, with the
responsible tap hooked in as the driving tap.  Taps at existing
paths become diverted to the driving taps, enabling potential for
external drivers for tapped variables.

This commit only adds the implementation, nothing is actually
using it yet.