main,til_args: employ the stream, add --print-pipes
This is a rudimentary integration of the new til_stream_t into
rototiller.  If the stream is going to continue living in
til_fb_fragment_t, the fragmenters and other nested frame
scenarios likely need to be updated to copy the stream through to
make the pipes available to the nested renders.

--print-pipes dumps the values found at the pipes' driver taps
to stdout on every frame.

Right now there's no way to externally write these values, but
with --print-pipes you can already see where things are going and
it's a nice visibility tool for tapped variables in modules.

Only stars and plato tap variables presently, but that will
improve.