* turn til_fb_fragment_t.stream into a discrete parameter
This was mostly done out of convenience at the expense of turning
the fragment struct into more of a junk drawer.

But properly cleaning up owned stream pipes on context destroy
makes the inappropriateness of being part of til_fb_fragment_t
glaringly apparent.

Now the stream is just a separate thing passed to context create,
with a reference kept in the context for use throughout.  Cleanup
of the owned pipes on the stream supplied to context create is
automagic when the context gets destroyed.

Note that despite there being a stream in the module context, the
stream to use is still supplied to all the rendering family
functions (prepare/render/finish) and it's the passed-in stream
which should be used by these functions.  This is done to support
the possibility of switching out the stream frame-to-frame, which
may be interesting.  Imagine doing things like a latent stream
and a future stream and switching between them on the fly for
instance.  If there's a sequencing composite module, it could
flip between multiple sets of tracks or jump around multiple
streams with the visuals immediately flipping accordingly.

This should fix the --print-pipes crashing issues caused by lack
of cleanup when contexts were removed (like rtv does so often).