modules/compose: fix segfault introduced by 83e41d
It was assumed (n_modules - n_overlayable) would give the number
of non-overlayable modules appropriate as base layers.  But with
the skipping of hermetic and experimental modules the base_idx
could be out of reach leaving layers NULL after the loop, which
will segfault later when strlen() assumes it's non-NULL.

This commit does the simple thing and also counts the unusable
modules to subtract from those eligible for base layers along
with n_overlayable.