til_stream: add a second void* to til_stream_pipe_t
It seems likely that pipe owners will need not only a way to
differentiate themselves via the owner pointer, but also
somewhere to register a pipe-specific reference.

There probably needs to be a result pointer added for storing the
owner_foo when the owner taps, so the owner can make use of it.