til_stream: introduce stream iterator et al
In order to implement something like a rocket module there needs
to be a way to iterate the pipes in the stream, and take
owernship of them when not already owned by rocket.

The way rocket's API works is you lookup tracks by name at
runtime.  The rocket module will be a meta module that calls into
another module for rendering, arbitrarily configured via a rocket
setting a la checkers::fill_module.

So it won't be until the underlying modules do some rendering
that their taps get their respective pipes established in the
stream.  Then the rocket module can look at all the pipes and for
any it doesn't own yet, it can get the tracks for those names and
take ownership while stowing the track handle in owner_foo for
the pipe.

While iterating all the pipes, the pipes already owned will have
the tracks readily available which can produce the values to
stick in the tap.

Something like that anyways, the til_stream_t api changes in this
commit are all preparatory for a rocket module.