til_fb: don't dereference NULL fragment ops
For strictly logical fragments (e.g. tiled fragmenters) there
won't be any ops, and that's even documented in the comments.

But the snapshot and reclaim functoins were assuming the ops
would be non-NULL.  Snapshot in particular trips on this
assumption when a module snapshots a subfragment, like drizzle in
montage.  I'm surprised I haven't encountered this crash
before...