til_stream: teardown pipes when the driving_tap's owner matches
The driving tap's owner and pipe's owner are decoupled.  When
tearing down an owner from a stream, any pipes its taps are
driving should also just go away.  Otherwise its taps could
linger on pipes it doesn't own, which would be a UAF bug.

If the pipe is still needed, it'll just get recreated by another
tap.  So there's a small perf hit, but this shouldn't be a
continuos kind of occurrence.