modules/drizzle: add taps for viscosity,rainfall
Just some obvious taps...

Note the actual usable range for viscosity is quite small, like
no greater than .05 really works.  There still needs to be some
way to describe bounds on the taps, that or normalizing things to
always be 0-1 or -1..+1 kind of thing, and expecting the modules
or the tap api to map those to the sane ranges.

Just leave everything raw for now, wiring up the taps at least
opens up experimentation and getting a feel for what makes most
sense.