modules/rocket: %s/rocket/rkt/g
It's annoying to have the til module called rocket, and the sync
tracker protocol/library called rocket, so let's at least
differentiate it in code/comments/textual discussion.

Plus this results in shorter module context paths i.e.:
/rkt:scene
/rkt/compose/drizzle:rain
/rkt/compose/drizzle:viscosity
/rkt/compose/plato:spin_rate
/rkt/compose/plato:orbit_rate

vs.

/rocket:scene
etc...

These names are shown in the editor, and they'll tend to be long
but let's at least get the root name down to three chars this
way.

A rename of the files and build system update will come in a
subsequent commit